requirements.txt
fastapi
uvicorn
gunicorn
psycopg2-binary
prettytable
httpx